+++ src/Document/Library/Bridge/Symfony/DependencyInjection/ZenstruckDocumentLibraryExtension.php
@@ -17,9 +17,9 @@ use Zenstruck\Document\Library\Bridge\Doctrine\Persistence\MappingProvider;
 use Zenstruck\Document\Library\Bridge\Symfony\Form\DocumentType;
 use Zenstruck\Document\Library\Bridge\Symfony\Form\PendingDocumentType;
 use Zenstruck\Document\Library\Bridge\Symfony\HttpKernel\DoctrineMappingProviderCacheWarmer;
+use Zenstruck\Document\Library\Bridge\Symfony\HttpKernel\PendingDocumentValueResolver;
+use Zenstruck\Document\Library\Bridge\Symfony\HttpKernel\RequestFilesExtractor;
 use Zenstruck\Document\Library\Bridge\Symfony\Serializer\LazyDocumentNormalizer;
-use Zenstruck\Document\Library\Bridge\Symfony\ValueResolver\PendingDocumentValueResolver;
-use Zenstruck\Document\Library\Bridge\Symfony\ValueResolver\RequestFilesExtractor;
 use Zenstruck\Document\Library\FlysystemLibrary;
 use Zenstruck\Document\LibraryRegistry;
 use Zenstruck\Document\Namer;
